Sponsor y sponsorship
Falta arreglar creditcard y la query que muestre solo las parades aceptadas

+++ D02 - Functional Testing/Item 3 - Project/Mari/Acme-Parade/src/main/java/services/ParadeService.java
@@ -62,6 +62,8 @@ public class ParadeService {
 		p.setFinalMode(false);
 		p.setRequests(new ArrayList<Request>());
 		p.setFloats(new ArrayList<Float>());
+		p.setStatus("");
+		p.setWhyRejected("");
 		return p;
 	}
 	public Parade save(final Parade parade) {
@@ -91,9 +93,11 @@ public class ParadeService {
 		b = this.processionRepository.findBrotherhoodByParadesId(idParade);
 		Collection<Parade> paradesPerBrotherhood;
 		paradesPerBrotherhood = b.getParades();
-		paradesPerBrotherhood.remove(p);
-		b.setParades(paradesPerBrotherhood);
-		this.processionRepository.delete(p);
+		if (paradesPerBrotherhood.contains(p)) {
+			paradesPerBrotherhood.remove(p);
+			b.setParades(paradesPerBrotherhood);
+			this.processionRepository.delete(p);
+		}
 	}
 
 	public Parade reconstruct(final Parade parade, final BindingResult binding) {
@@ -109,8 +113,42 @@ public class ParadeService {
 			result.setFinalMode(parade.getFinalMode());
 			result.setFloats(parade.getFloats());
 			result.setRequests(parade.getRequests());
+			result.setWhyRejected(parade.getWhyRejected());
+			result.setStatus(parade.getStatus());
+
 		}
 		this.validator.validate(result, binding);
 		return result;
+	}
+
+	public Parade copyParade(final int id) {
+
+		Parade toCopy;
+		toCopy = this.processionRepository.findOne(id);
+
+		Parade newParade;
+		newParade = this.createParade();
+
+		newParade.setDescription(toCopy.getDescription());
+		newParade.setFinalMode(false);
+		newParade.setFloats(toCopy.getFloats());
+		newParade.setMomentOrganised(toCopy.getMomentOrganised());
+		newParade.setRequests(toCopy.getRequests());
+		newParade.setTicker(Utiles.generateTicker());
+		newParade.setStatus("");
+		newParade.setWhyRejected("");
+
+		Parade saved;
+		saved = this.processionRepository.save(newParade);
+
+		Brotherhood b;
+		b = this.findBrotherhoodByParade(toCopy.getId());
+
+		Collection<Parade> parades;
+		parades = b.getParades();
+		parades.add(saved);
+		b.setParades(parades);
+
+		return saved;
 	}
 }
